Removed JsonApiInputFormatter workaround
